feat: Vault granularity (#234)

+++ test/fuzz/DistributionVaultFuzz.t.sol
@@ -4,7 +4,6 @@ pragma solidity 0.8.23;
 
 import { PureEpochs } from "../../src/libs/PureEpochs.sol";
 
-import { IDistributionVault } from "../../src/interfaces/IDistributionVault.sol";
 import { DistributionVault } from "../../src/DistributionVault.sol";
 
 import { MockERC20, MockEpochBasedVoteToken } from "./../utils/Mocks.sol";
@@ -36,13 +35,14 @@ contract DistributionVaultTests is TestUtils {
         uint256 account2PastBalance,
         uint256 pastTotalSupply
     ) external {
-        token1Balance = bound(token1Balance, 0, type(uint128).max);
-        account1PastBalance = bound(account1PastBalance, 0, type(uint128).max); //@elcid closed environment.
-        account2PastBalance = bound(account2PastBalance, 0, type(uint128).max);
+        token1Balance = bound(token1Balance, 0, type(uint112).max);
+        account1PastBalance = bound(account1PastBalance, 0, type(uint112).max); //@elcid closed environment.
+        account2PastBalance = bound(account2PastBalance, 0, type(uint112).max);
 
-        vm.assume(account1PastBalance + account2PastBalance < type(uint128).max);
+        vm.assume(account1PastBalance + account2PastBalance < type(uint112).max);
         vm.assume(account1PastBalance + account2PastBalance > 0);
-        pastTotalSupply = bound(pastTotalSupply, account1PastBalance + account2PastBalance, type(uint128).max);
+
+        pastTotalSupply = bound(pastTotalSupply, account1PastBalance + account2PastBalance, type(uint112).max);
 
         _baseToken.setPastBalanceOf(_accounts[0], PureEpochs.currentEpoch(), account1PastBalance); //@elcid sets ZERO balance
         _baseToken.setPastBalanceOf(_accounts[1], PureEpochs.currentEpoch(), account2PastBalance);
@@ -50,6 +50,7 @@ contract DistributionVaultTests is TestUtils {
 
         _token1.setBalance(address(_vault), token1Balance); //@elcid sets token.balanceOf
         _vault.distribute(address(_token1));
+
         assertEq(_vault.distributionOfAt(address(_token1), PureEpochs.currentEpoch()), token1Balance);
 
         uint256 startEpoch_ = PureEpochs.currentEpoch();
@@ -59,6 +60,7 @@ contract DistributionVaultTests is TestUtils {
 
         vm.prank(_accounts[0]);
         _vault.claim(address(_token1), startEpoch_, endEpoch_, _accounts[0]);
+
         vm.prank(_accounts[1]);
         _vault.claim(address(_token1), startEpoch_, endEpoch_, _accounts[1]);
     }
